Create lowerspeed.py
Replace speed's value win the speed with ' around it. Replace username's value with username with ' around it. Replace password's value with password with ' around it. This will lower your speed so that you can race much faster. Also, can keep selling the car given to you to get free money.